server/pyproject.toml
[tool.poetry]
name = "volview_server"
version = "0.1.0"
description = "The VolView Python Server"
authors = ["Forrest <forrest.li@kitware.com>", "Andrew Howe <andrew.howe@kitware.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
itk = "^5.3.0"
black = "^23.1.0"
flake8 = "^6.0.0"
pytest = "^7.2.1"
numpy = "^2.3.3"
aiohttp = "3.9.2"
python-socketio = "^5.8.0"
charset-normalizer = "3.4.3"
setuptools = "80.9.0"
certifi = "2025.8.3"
filelock = "3.19.1"
fire = "0.7.1"
fsspec = "<2025.0.0"
hf-xet = "1.1.10"
huggingface-hub = "0.35.3"
idna = "3.10"
ignite = "1.1.0"
jinja2 = "3.1.6"
markupsafe = "3.0.3"
monai = "1.5.1"
mpmath = "1.3.0"
networkx = "3.5"
nibabel = "5.3.2"
packaging = ">=20.0,<25.0"
pynrrd = "1.1.3"
pytorch-ignite = "0.5.2"
pyyaml = "6.0.3"
requests = "2.32.5"
sympy = "1.14.0"
termcolor = "3.1.0"
torch = "2.8.0"
torchvision = "0.23"
lightning = "2.1.3"
tqdm = "4.67.1"
typing-extensions = "4.15.0"
urllib3 = "2.5.0"
wrapt = "1.17.3"
pydicom = "2.4.4"
einops = "0.7.0"
scikit-learn = "^1.7.0"
transformers = ">=4.50.0"
accelerate = "1.9.0"
Pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.22.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
